requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
joblib==0.13.1
MarkupSafe==1.1.0
numpy==1.16.1
pymongo==3.7.2
scikit-surprise==1.0.6
scipy==1.2.1
six==1.12.0
Werkzeug==0.14.1
